Created Sales List view and Sales Creation form on the frontend. Created Nav links and Routed as well.

+++ sales/api/sales_rest/models.py
@@ -34,4 +34,4 @@ class Sale(models.Model):
         related_name="sales",
         on_delete=models.CASCADE,
     )
-    price = models.PositiveBigIntegerField()
+    price = models.FloatField()
